feat (prefabs): move configs from interactables
