+++ include/game/prefabs/config/weapon_bat_config.h
